requirements.txt
pulp
networkx
fastapi
uvicorn[standard]
jinja2
python-multipart
aiofiles
numpy

# For auth 
python-dotenv 
pydantic[email]
passlib[bcrypt] 
bcrypt==3.1.7
python-jose[cryptography] 
#jwt


# For testing
pytest
httpx
pytest-cov
